requirements.txt
streamlit==1.31.0
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
concurrent-log-handler==0.9.24
pyppeteer==1.0.2
httpx==0.25.2
python-dotenv==1.0.0
readability-lxml==0.8.1
